ataraxia/ai-sdk/requirements.txt
qiniu==7.1.4
pytest==3.1.3
requests==2.18.1
pytest-cov==2.5.1
deepdiff==3.3.0
yapf==0.24.0
pylint==1.9.3
grpcio==1.14.2
grpcio-tools==1.14.2
pyzmq==17.1.2
ipdb==0.11
schema==0.6.8
jinja2==2.10
docopt==0.6.2
pytest-benchmark==3.1.1
opencv-python==4.0.0.21
prometheus_client==0.4.0
six==1.11.0
